finished the protection of crud operation through jwt tokens

+++ controllers/contactController.js
@@ -2,15 +2,15 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@description Get all contacts
 //@route GET /api/contacts
-//@access Public
+//@access Private
 const getContacts =asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({user_id: req.user.id});
   res.status(200).json(contacts);
 }) ;
 
 //@description Create New contacts
 //@route POST /api/contacts
-//@access Public
+//@access Private
 const createContact = asyncHandler (async (req, res) => {
     console.log("The request Body is :", req.body)
     const{name, email, phone} = req.body;
@@ -22,13 +22,14 @@ const createContact = asyncHandler (async (req, res) => {
       name,
       email,
       phone,
+      user_id: req.user.id,
     });
   res.status(201).json(contact);
 });
 
 //@description GET Individual contacts
 //@route GET /api/contacts/:id
-//@access Public
+//@access Private
 const getContact =asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if(!contact){
@@ -40,12 +41,17 @@ const getContact =asyncHandler(async (req, res) => {
 
 //@description Update contacts
 //@route PUT /api/contacts/:id
-//@access Public
+//@access Private
 const updateContact =asyncHandler (async (req, res) => {
    const contact = await Contact.findById(req.params.id);
   if(!contact){
     res.status(404);
     throw new Error ("Contact not found");
+  }
+
+  if(contact.user_id.toString()!== req.user.id){
+    res.status(403);
+    throw new Error("User Donot have permission to update this contact");
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -59,12 +65,16 @@ const updateContact =asyncHandler (async (req, res) => {
 
 //@description Delete contacts
 //@route Delete /api/contacts/:id
-//@access Public
+//@access Private
 const deleteContact = asyncHandler(async (req, res) => {
    const contact = await Contact.findById(req.params.id);
   if(!contact){
     res.status(404);
     throw new Error ("Contact not found");
+  }
+  if(contact.user_id.toString()!== req.user.id){
+    res.status(403);
+    throw new Error("User Donot have permission to update this contact");
   }
   await Contact.deleteOne({ _id: req.params.id });
 //adding comment to check the commmit
